tools/env: Add include of stdint.h when MTD_OLD is not defined.
The source file fw_env.c uses the types uint8_t and uint32_t frequently.
It indirectly accesses their definition through the include of the header
mtd/mtd-user.h, which in turn includes the header stdint.h, where the
above types are defined. In the newer kernel versions, though, the header
mtd/mtd-user.h does not include the header stdint.h anymore, so the
compiler does not recognize the uint8_t and uint32_t types.
The following commit moves up the include of stdint.h to affect cases
where MTD_OLD is not defined to fix this issue.

Change-Id: I69b27f2f6ad1323777c108091ad2f3f6f03a2af5
Signed-off-by: Liav Rehana <liavr@mellanox.com>