board/ezchip: Add indication of actual PCI gen being used during start-up.
When starting up the PCI link, the NPS chip sometimes does not acquire
the PCI link speed that it requested due to hardware limitation.
In these cases, the PCI link will be established, but with a lower
frequency, which will be unknown to users who won't check that elsewhere.
The following commit provides an indication to the actual PCI gen that
was established during start-up.

Change-Id: Ieeaf1f33274bff92e61661428708e74df417dc72
Signed-off-by: Liav Rehana <liavr@mellanox.com>