board/ezchip: Fix trigger of error response.
Sometimes error response is triggered during the polling of the link
status register of the PEXC block. This is probably due to the fact that
there are read requests from the PEXC block when the two PCI end-points
are not synchronized, which is during the change of the PCI link speed.
The following commit fixes that trigger by polling the link speed first,
and only when it is updated do we perform polling of the link status in
the PEXC block.

Change-Id: Ie502afffa48857e3dd698326aa6a0ef2e9955d90
Signed-off-by: Liav Rehana <liavr@mellanox.com>